requirements.txt
pytest
allure-pytest
requests
jsonschema
